Fix blob driver.GetTable table existence check query
Crate renamed the schema_name column to table_schema to be SQL
compliant, as of:

https://github.com/crate/crate/pull/4218

PS:

I'm not even sure we should try to query for a table existence check,
since:

a) This might not be a expected behaviour when initializing a new "table
driver" through driver.GetTable. We could provide another API for that.
b) Crate's information_schema is not "stable api", it will change from
version to versio.